refactor(events): Refactor event file into multiple files
