+++ src/events/event-model.ts
@@ -1,19 +1,13 @@
 /**
- * Model for game events in the Regal Game Library.
+ * Interfaces for the different types of events, and functions for manipulating them.
+ *
  * @since 0.3.0
  * @author Joe Cowman
  * @license MIT (see https://github.com/regal/regal)
  */
 
-import { PropertyChange, PropertyOperation } from "./agent";
-import { RegalError } from "./error";
-import GameInstance from "./game-instance";
-import { OutputLine } from "./output";
-
-/** Event ID for untracked EventFunctions. */
-export const DEFAULT_EVENT_ID: number = 0;
-/** Name of untracked EventFunctions. */
-export const DEFAULT_EVENT_NAME: string = "DEFAULT";
+import { RegalError } from "../error";
+import GameInstance from "../game-instance";
 
 /**
  * A function that modifies the game instance.
@@ -98,206 +92,6 @@ export const noop: TrackedEvent = (() => {
 
     return event;
 })();
-
-/**
- * Record of a TrackedEvent's effects in a game cycle.
- */
-export class EventRecord {
-    /** Default EventRecord for untracked EventFunctions. */
-    public static default = new EventRecord();
-    /** The IDs of the OutputLines emitted by the event. */
-    public output?: number[];
-    /** The ID of the event that caused this event. */
-    public causedBy?: number;
-    /** The IDs of the events that were caused by this event. */
-    public caused?: number[];
-    /** The records of all changes to registered agents that were caused by this event. */
-    public changes?: PropertyChange[];
-
-    /**
-     * Constructs a new EventRecord.
-     * @param id The event's unique numeric ID.
-     * @param name The event's name.
-     * @param func The event's TrackedEvent.
-     */
-    constructor(
-        public id: number = DEFAULT_EVENT_ID,
-        public name: string = DEFAULT_EVENT_NAME,
-        public func: TrackedEvent = noop
-    ) {}
-
-    /**
-     * Appends a reference to an OutputLine to the EventRecord's output log.
-     * @param line The line of output emitted by the event.
-     */
-    public trackOutputWrite(line: OutputLine): void {
-        if (this.output === undefined) {
-            this.output = [];
-        }
-        this.output.push(line.id);
-    }
-
-    /**
-     * Creates an association between this EventRecord and the caused events,
-     * setting the caused events' causedBy properties to this EventRecord's ID, and
-     * appending the events' IDs to this EventRecord's caused property.
-     * @param events The events that were caused by this event.
-     */
-    public trackCausedEvent(...events: EventRecord[]): void {
-        if (this.caused === undefined) {
-            this.caused = [];
-        }
-        this.caused.push(...events.map(e => e.id));
-        events.forEach(e => (e.causedBy = this.id));
-    }
-
-    /**
-     * Adds a record of a single change to a registered agent's property to the
-     * EventRecord's changes property.
-     * @param agentId The registered agent's ID.
-     * @param property The property being modified.
-     * @param op The type of modification.
-     * @param init The initial value of the property.
-     * @param final The final value of the property.
-     */
-    public trackChange<T>(
-        agentId: number,
-        property: PropertyKey,
-        op: PropertyOperation,
-        init?: T,
-        final?: T
-    ): void {
-        if (this.changes === undefined) {
-            this.changes = [];
-        }
-        this.changes.push({
-            agentId,
-            final,
-            init,
-            op,
-            property: property.toString()
-        });
-    }
-}
-
-/**
- * Manager for all events in a GameInstance.
- */
-export class InstanceEvents {
-    /** Contains records of the past events executed during the game cycle. */
-    public history: EventRecord[] = [];
-
-    /** Internal member for the ID of the most recently generated EventRecord. */
-    private _lastEventId;
-    /** Internal queue of events that have yet to be executed. */
-    private _queue: EventRecord[] = [];
-
-    /**
-     * Constructs an InstanceEvents.
-     * @param game The game instance that owns this InstanceEvents.
-     * @param startingEventId Optional starting ID for new `EventRecord`s.
-     */
-    constructor(public game: GameInstance, startingEventId = DEFAULT_EVENT_ID) {
-        this._lastEventId = startingEventId;
-    }
-
-    /** The current EventRecord. */
-    get current(): EventRecord {
-        let event = this._queue[0];
-
-        if (event === undefined) {
-            event = EventRecord.default;
-        }
-
-        return event;
-    }
-
-    /** The ID of the most recently generated `EventRecord`. */
-    get lastEventId() {
-        return this._lastEventId;
-    }
-
-    /**
-     * Executes the given event and all events caused by it.
-     * @param event The TrackedEvent to be invoked.
-     */
-    public invoke(event: TrackedEvent): void {
-        this._addEvent(event);
-        this._executeCurrent();
-    }
-
-    /**
-     * Adds the event to the internal queue. If the event is an EventQueue,
-     * the event's immediateEvents are added to the front of the queue
-     * and the delayedEvents are added to the back of the queue.
-     * @param event The TrackedEvent to be added to the queue.
-     * @param cause The EventRecord to be recorded as the event's cause (optional).
-     */
-    public _addEvent(event: TrackedEvent, cause?: EventRecord): void {
-        let immediateEvents: TrackedEvent[];
-        let delayedEvents: TrackedEvent[];
-
-        if (isEventQueue(event)) {
-            immediateEvents = event.immediateEvents;
-            delayedEvents = event.delayedEvents;
-        } else {
-            immediateEvents = [event];
-            delayedEvents = [];
-        }
-
-        const mapToRecord = (event: TrackedEvent) =>
-            new EventRecord(++this._lastEventId, event.eventName, event);
-
-        const immediateEventRecords = immediateEvents.map(mapToRecord);
-        const delayedEventRecords = delayedEvents.map(mapToRecord);
-
-        if (cause) {
-            cause.trackCausedEvent(...immediateEventRecords);
-            cause.trackCausedEvent(...delayedEventRecords);
-        }
-
-        this._queue = immediateEventRecords.concat(this._queue);
-        this._queue = this._queue.concat(delayedEventRecords);
-    }
-
-    /**
-     * Executes the current EventRecord and recursively exeuctes
-     * all remaining EventRecords in the queue.
-     */
-    public _executeCurrent(): void {
-        const current = this.current;
-        const nextEvent = current.func.target(this.game);
-        this._archiveCurrent();
-
-        // Add the nextEvent to the internal queue, if necessary.
-        if (isTrackedEvent(nextEvent) && nextEvent !== noop) {
-            this._addEvent(nextEvent, current);
-        }
-
-        // While the queue is not empty, keep executing.
-        if (this._queue.length > 0) {
-            this._executeCurrent();
-        }
-    }
-
-    /**
-     * Deletes unnecessary data from the current EventRecord
-     * and moves it to the history array.
-     */
-    public _archiveCurrent(): void {
-        delete this.current.func;
-        this.history.unshift(this._queue.shift());
-    }
-
-    /**
-     * Creates a new `InstanceEvents` for the new game cycle.
-     * **Don't call this unless you know what you're doing.**
-     * @param current The `GameInstance` for the new game cycle.
-     */
-    public cycle(current: GameInstance): InstanceEvents {
-        return new InstanceEvents(current, this.lastEventId);
-    }
-}
 
 /** Creates a function that returns an error upon invocation. */
 const illegalEventQueueInvocation = () => (game: GameInstance): undefined => {
