+++ test/api-hooks.test.ts
@@ -8,7 +8,7 @@ import {
     onBeforeUndoCommand
 } from "../src/api-hooks";
 import GameInstance from "../src/game-instance";
-import { noop, on } from "../src/event";
+import { noop, on } from "../src/events";
 import { log } from "./utils";
 import { PropertyOperation } from "../src/agent";
 import { RegalError } from "../src/error";
